pyproject.toml
[tool.poetry]
name = "hr_policy_chatbot"
version = "0.1.0"
description = "Financial Institution HR Policy Chatbot using RAG and Gemini LLM"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
google-generativeai = "0.3.1"
scikit-learn = "^1.3.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
nltk = "^3.8.1"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "hr_policy_chatbot.main:main"
